+++ test/libsolidity/syntaxTests/conversion/conversion_to_bytes.sol
@@ -1,0 +1,5 @@
+contract test {
+  function f() public pure returns (bytes) {
+    return bytes("abc");
+  }
+}